feat: split test in two to prevent a juju bug
