+++ tests/integration/upgrade/test_sharding_rollback.py
@@ -2,13 +2,12 @@
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-import asyncio
 from pathlib import Path
 
 import pytest
 from pytest_operator.plugin import OpsTest
 
-from ..helpers import DEPLOYMENT_TIMEOUT, find_unit, wait_for_mongodb_units_blocked
+from ..helpers import DEPLOYMENT_TIMEOUT, find_unit
 from ..sharding_tests.helpers import deploy_cluster_components, integrate_cluster
 from ..sharding_tests.writes_helpers import (
     SHARD_ONE_DB_NAME,
@@ -118,76 +117,6 @@ async def test_rollback_on_config_server(
     ), "continuous writes to shard two failed during upgrade"
 
     await ops_test.model.wait_for_idle(apps=CLUSTER_COMPONENTS, timeout=1000, idle_period=20)
-
-    # after all shards have upgraded, verify that the balancer has been turned back on
-    # TODO implement this check once we have implemented the post-cluster-upgrade code DPE-4143
-
-
-@pytest.mark.runner(["self-hosted", "linux", "X64", "jammy", "large"])
-@pytest.mark.group(1)
-@pytest.mark.abort_on_fail
-async def test_rollback_on_shard_and_config_server(
-    ops_test: OpsTest, continuous_writes_to_shard_one, continuous_writes_to_shard_two
-) -> None:
-    """Verify that a config-server and shard can safely rollback without losing writes."""
-    new_charm = await ops_test.build_charm(".")
-    await run_upgrade_sequence(ops_test, CONFIG_SERVER_APP_NAME, new_charm=new_charm)
-
-    with open("charm_version", mode="r") as fd:
-        revision = fd.read().strip()
-
-    # Wait for statuses to settle down
-    asyncio.gather(
-        wait_for_mongodb_units_blocked(ops_test, SHARD_ONE_APP_NAME),
-        wait_for_mongodb_units_blocked(ops_test, SHARD_TWO_APP_NAME),
-        ops_test.model.wait_for_idle(
-            apps=[CONFIG_SERVER_APP_NAME],
-            timeout=1000,
-            idle_period=20,
-            status=f"Waiting for shards to upgrade/downgrade to revision {revision}-locally built.",
-        ),
-    )
-
-    await run_upgrade_sequence(ops_test, SHARD_ONE_APP_NAME, new_charm=new_charm)
-
-    # Wait for statuses to settle down
-    asyncio.gather(
-        wait_for_mongodb_units_blocked(ops_test, SHARD_TWO_APP_NAME),
-        ops_test.model.wait_for_idle(apps=[SHARD_ONE_APP_NAME], timeout=1000, idle_period=20),
-        ops_test.model.wait_for_idle(
-            apps=[CONFIG_SERVER_APP_NAME],
-            timeout=1000,
-            idle_period=20,
-            status=f"Waiting for shards to upgrade/downgrade to revision {revision}-locally built.",
-        ),
-    )
-
-    await refresh_with_juju(ops_test, CONFIG_SERVER_APP_NAME, channel="6/edge")
-
-    # verify no writes were skipped during upgrade process
-    shard_one_expected_writes = await stop_continous_writes(
-        ops_test,
-        config_server_name=CONFIG_SERVER_APP_NAME,
-        db_name=SHARD_ONE_DB_NAME,
-    )
-    shard_two_expected_writes = await stop_continous_writes(
-        ops_test,
-        config_server_name=CONFIG_SERVER_APP_NAME,
-        db_name=SHARD_TWO_DB_NAME,
-    )
-
-    shard_one_actual_writes = await count_shard_writes(
-        ops_test, CONFIG_SERVER_APP_NAME, SHARD_ONE_DB_NAME
-    )
-    shard_two_actual_writes = await count_shard_writes(
-        ops_test, CONFIG_SERVER_APP_NAME, SHARD_TWO_DB_NAME
-    )
-    assert (
-        shard_one_actual_writes >= shard_one_expected_writes["number"]
-    ), "continuous writes to shard one failed during upgrade"
-    assert (
-        shard_two_actual_writes >= shard_two_expected_writes["number"]
-    ), "continuous writes to shard two failed during upgrade"
 
     # after all shards have upgraded, verify that the balancer has been turned back on
     # TODO implement this check once we have implemented the post-cluster-upgrade code DPE-4143
